269 prihlaseni aktivniho uzivatele jako ucastnika na udalost

+++ one_time_events/models.py
@@ -11,6 +11,7 @@ from events.models import (
     OrganizerPositionAssignment,
 )
 from trainings.models import Training
+from transactions.models import Transaction
 from vzs import settings
 
 
@@ -43,6 +44,44 @@ class OneTimeEvent(Event):
 
     state = models.CharField(max_length=10, choices=EventOrOccurrenceState.choices)
 
+    def does_participant_satisfy_requirements(self, person):
+        if not super().does_participant_satisfy_requirements(person):
+            return False
+
+        if (
+            self.training_category is not None
+            and not Training.does_person_attends_training_of_category(
+                person, self.training_category
+            )
+        ):
+            return False
+
+        return True
+
+    def has_free_spot(self):
+        if self.participants_enroll_state == ParticipantEnrollment.State.APPROVED:
+            return len(self.approved_participants()) < self.capacity
+        elif self.participants_enroll_state == ParticipantEnrollment.State.SUBSTITUTE:
+            return len(self.all_possible_participants()) < self.capacity
+        raise NotImplementedError
+
+    def can_participant_unenroll(self, person):
+        if not super().can_participant_unenroll(person):
+            return False
+
+        enrollment = self.get_participant_enrollment(person)
+        if enrollment.transaction is None:
+            return True
+        return not enrollment.transaction.is_settled
+
+    def get_participant_enrollment(self, person):
+        if person is None:
+            return None
+        try:
+            return person.onetimeeventparticipantenrollment_set.get(one_time_event=self)
+        except OneTimeEventParticipantEnrollment.DoesNotExist:
+            return None
+
     def _occurrences_list(self):
         return OneTimeEventOccurrence.objects.filter(event=self)
 
@@ -54,15 +93,8 @@ class OneTimeEvent(Event):
         occurrences = self._occurrences_list().order_by("date")
         return occurrences
 
-    def enrollments_by_state(self, state):
-        output = []
-        for enrolled_participant in self.enrolled_participants.all():
-            enrollment = enrolled_participant.onetimeeventparticipantenrollment_set.get(
-                one_time_event=self
-            )
-            if enrollment.state == state:
-                output.append(enrollment)
-        return output
+    def enrollments_by_Q(self, condition):
+        return self.onetimeeventparticipantenrollment_set.filter(condition)
 
     def __str__(self):
         return self.name
@@ -99,13 +131,31 @@ class OneTimeEventParticipantEnrollment(ParticipantEnrollment):
         "transactions.Transaction", null=True, on_delete=models.SET_NULL
     )
 
+    class Meta:
+        unique_together = ["one_time_event", "person"]
+
     def delete(self):
         transaction = OneTimeEventParticipantEnrollment.objects.get(
             pk=self.pk
         ).transaction
         super().delete()
-        if not transaction.is_settled:
+        if transaction is not None and not transaction.is_settled:
             transaction.delete()
+
+    @staticmethod
+    def create_attached_transaction(enrollment, event):
+        fee = (
+            -enrollment.agreed_participation_fee
+            if enrollment.agreed_participation_fee is not None
+            else -event.default_participation_fee
+        )
+        return Transaction(
+            amount=fee,
+            reason=f"Schválená přihláška na jednorázovou událost {event}",
+            date_due=event.date_start,
+            person=enrollment.person,
+            event=event,
+        )
 
     @property
     def event(self):
